requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy[asyncio]
alembic
pydantic
python-dotenv
